Entity page

+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityLocationsHandler.java
@@ -75,8 +75,8 @@ public class UserActivityLocationsHandler extends UserActivityBaseHandler {
     }
 
     @Override
-    Function<Integer, Integer> valueReducer() {
-        return (newValue) -> 1;
+    Function<Double, Double> valueReducer() {
+        return (newValue) -> 1.0;
     };
 
     @Override
